fix: UI audit fixes — orphan outputs, missing inputs, CSS gaps
- Add missing download_worldclim checkbox to sidebar (was dead code branch)
- Fix as.integer truncation of gd_worldclim_res 2.5 → as.numeric
- Remove orphan suitability_plot renderPlot (no UI element uses it)
- Add CSS for flagged-count-badge and flagged-actions flex container
- Change run_log from verbatimTextOutput to htmlOutput to match renderUI
- Add CSS for .flagged-actions layout container

+++ R/mod_get_data.R
@@ -132,7 +132,7 @@ mod_get_data_server <- function(id, rv, input) {
 
     observeEvent(input$gd_download_worldclim, {
       source <- input$gd_climate_source %||% "worldclim"
-      res <- as.integer(input$gd_worldclim_res %||% 10)
+      res <- as.numeric(input$gd_worldclim_res %||% 10)
       gd_append_log("gd_worldclim_log", paste0("Starting WorldClim download (source=", source, ", res=", res, ")..."))
       download_covariate_bg(
         log_target = "gd_worldclim_log", log_append = gd_append_log,
